Refactor: Moved InteractionObjectTestBuilder to the testing module
