+++ testing/build.gradle
@@ -102,6 +102,7 @@ dependencies {
       'com.google.truth:truth:1.1.3',
       'junit:junit:4.12',
       "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version",
+      "org.jetbrains.kotlin:kotlin-test",
       'org.mockito:mockito-core:2.19.0',
       project(":domain"),
   )
@@ -114,6 +115,7 @@ dependencies {
   annotationProcessor(
       'com.google.auto.service:auto-service:1.0-rc4',
   )
+  implementation project(":domain")
   // TODO(#59): Remove this once Bazel is introduced
   // sufficiently visible for generated Dagger code. This can be done more cleanly via Bazel since dependencies can be
   // controlled more directly than in Gradle.
