+++ testing/src/main/java/org/oppia/android/testing/InteractionObjectTestBuilder.kt
@@ -1,5 +1,5 @@
 
-package org.oppia.android.domain.classify
+package org.oppia.android.testing
 
 import org.oppia.android.app.model.Fraction
 import org.oppia.android.app.model.HtmlTranslationList
